UI changes for location bar, Refactoring of inbox
