+++ frontend/public/lib/communicationOperations.js
@@ -1,3 +1,5 @@
+import { apiRequest } from "./apiOperations";
+
 // Returns the new comments object after adding a comment locally
 export function getCommentsObjectAfterAddingComment(c, comments) {
   //If the comment is a reply to a parent comment, just add it to the replies
@@ -34,4 +36,49 @@ export function getCommentsObjectAfterAddingComments(commentsToAdd, comments) {
     commentsObject = getCommentsObjectAfterAddingComment(comment, commentsObject);
   }
   return commentsObject;
+}
+
+export function parseChats(chats, texts) {
+  return chats
+    ? chats.map((chat) => ({
+        ...chat,
+        chatting_partner:
+          chat.participants.length === 2 && chat.participants.find((p) => p.id !== chat.user.id),
+        unread_count: chat.unread_count,
+        content: chat.last_message ? chat.last_message.content : texts.chat_has_been_created,
+      }))
+    : [];
+}
+
+export function parseChatData(chats) {
+  return chats.map((c) => ({
+    ...c,
+    participants: c.participants.map((p) => ({
+      ...p.user_profile,
+      role: p.role,
+      created_at: p.created_at,
+    })),
+  }));
+}
+
+export function getParsedChats(response, texts) {
+  const parsedChatData = parseChatData(response.data.results);
+  const parsedChats = parseChats(parsedChatData, texts);
+  return parsedChats;
+}
+
+export async function getResponseFromAPI(locale, token, url) {
+  try {
+    return await apiRequest({
+      method: "get",
+      url: url,
+      token: token,
+      locale: locale,
+    });
+  } catch (err) {
+    if (err.response && err.response.data) console.log("Error: " + err.response.data.detail);
+    console.log("error!");
+    console.log(err);
+    return null;
+  }
 }
